move naming helpers under compiler
