+++ src/compiler/codegen/sil/body.rs
@@ -7,12 +7,12 @@ use crate::compiler::model::{
     TemplateSelector, actor_enum_variant_const_expr, clause_actor_type_ref, observed_is_dynamic_binding, observed_open_bindings,
     observed_open_state_for_decl, parse_actor_enum_selector, parse_actor_enum_variant, spawn_target_state,
 };
+use crate::compiler::naming::{is_identifier, to_snake};
 use crate::compiler::syntax::body::{EntryBinding, EntryRoute, EntryStatement};
 use crate::compiler::syntax::lexer::{RESERVED_GENERATED_PREFIX, Span, Token, TokenKind, lex};
 use crate::compiler::syntax::word;
 use crate::compiler::syntax::*;
 use crate::error::{ArgentError, Result};
-use crate::naming::{is_identifier, to_snake};
 
 // Body lowering uses the surrounding Sil emitter's shared witness plans,
 // layout helpers, and generated names.
