+++ src/compiler/mod.rs
@@ -5,4 +5,5 @@
 pub(crate) mod codegen;
 pub(crate) mod loader;
 pub(crate) mod model;
+pub(crate) mod naming;
 pub(crate) mod syntax;
